flexgrid: renames
